requirements.txt
appdirs==1.4.3
Django==1.10.6
django-cors-headers==2.0.2
django-crispy-forms==1.6.1
django-markdown-deux==1.0.5
django-pagedown==0.1.1
djangorestframework==3.6.2
markdown2==2.3.3
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.2.0
six==1.10.0
